+++ scanapi/hide_utils.py
@@ -11,8 +11,8 @@ SENSITIVE_INFO_SUBSTITUTION_FLAG = "SENSITIVE_INFORMATION"
 def hide_sensitive_info(response):
     report_settings = settings.get("report", {})
     request = response.request
-    request_settings = report_settings.get("hide-request", {})
-    response_settings = report_settings.get("hide-response", {})
+    request_settings = report_settings.get("hide_request", {})
+    response_settings = report_settings.get("hide_response", {})
 
     _hide(request, request_settings)
     _hide(response, response_settings)