+++ scanapi/reporter.py
@@ -33,7 +33,7 @@ class Reporter:
     def _build_context(results):
         return {
             "now": datetime.datetime.now().replace(microsecond=0),
-            "project_name": settings.get("project-name", ""),
+            "project_name": settings.get("project_name", ""),
             "results": results,
             "session": session,
         }